+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 - `Service.DocumentService.CreateOneTimeDownloadLinkAsync` that allows to share document via one-time URL
 - `Service.UserService.CreateUserAsync` that allows to create an account for user
 - `Service.UserService.UpdateUserAsync` that allows to update user information i.e. first name, last name
+- `Service.UserService.SendVerificationEmailAsync` that allows to sends verification email to a user
 
 ## Changed
 - Added support for Basic Auth param to `Model.Token`